+++ src/crogger/logger.cc
@@ -28,7 +28,7 @@ namespace jowi::crogger {
     }
 
   public:
-    logger(size_t buf_size = 120) :
+    logger(uint64_t buf_size = 120) :
       __flt{std::make_unique<context_filter<no_filter>>()},
       __fmt{std::make_unique<formatter<colorful_formatter>>()},
       __emt{std::make_unique<stream_emitter<ss_emitter>>(buf_size, stdout_emitter{})} {}
@@ -45,7 +45,7 @@ namespace jowi::crogger {
       return *this;
     }
 
-    logger &set_emitter(is_emitter auto &&emt, size_t buf_size = 120) {
+    logger &set_emitter(is_emitter auto &&emt, uint64_t buf_size = 120) {
       __emt =
         std::make_unique<stream_emitter<ss_emitter>>(buf_size, std::forward<decltype(emt)>(emt));
       return *this;
